# shared/openapi/mesob-api.yaml
openapi: "3.1.0"
info:
  title: Mesob Wallet API
  version: "1.0.0"
  description: |
    Unified public API for the Mesob Wallet platform.
    All requests go through the API Gateway at port 8000.

    ## Authentication
    - Mobile/USSD users authenticate with **MSISDN + PIN** via `/v1/identity/login`
    - Access tokens are RS256 JWTs valid for **15 minutes**
    - Refresh tokens are 7-day opaque tokens returned as an `httpOnly` cookie
    - Staff users (ADMIN, BRANCH_MANAGER, SUPER_ADMIN) use the same auth flow

    ## Demo Accounts
    | Role | MSISDN | PIN |
    |------|--------|-----|
    | SUPER_ADMIN | +251911000001 | 111111 |
    | ADMIN | +251911000002 | 111111 |
    | BRANCH_MANAGER | +251911000003 | 111111 |
    | AGENT | +251911000004 | 111111 |
    | USER (KYC Tier 2) | +251911000005 | 111111 |
    | USER (KYC Tier 0) | +251911000006 | 111111 |

    ## Amounts
    All monetary values are in **ETB minor units (cents)**.
    Example: 100.00 ETB = `10000`

    ## Response Envelope
    All responses follow:
    ```json
    {
      "data": { ... },
      "meta": { "request_id": "...", "trace_id": "..." }
    }
    ```
    Errors:
    ```json
    {
      "error": { "code": "DOMAIN_ERROR_CODE", "message": "human message" },
      "meta": { ... }
    }
    ```
  contact:
    name: Mesob Engineering
    email: engineering@mesobwallet.et
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.mesobwallet.et/v1
...
paths:
...
  /branch/agents/{id}/approve:
    post:
      tags: [Branch]
      summary: Approve a pending agent (scoped to own region)
      operationId: approveAgent
      security:
        - bearerAuth: []
      description: Requires `BRANCH_MANAGER` role.
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Agent approved
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /branch/settlements:
    post:
      tags: [Branch]
      summary: Settle agent float (four-eyes gating)
      operationId: createSettlement
      security:
        - bearerAuth: []
      description: |
        Requires `BRANCH_MANAGER` role.
        Four-eyes: `second_authoriser_id` must differ from the caller's ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_id, amount_minor, second_authoriser_id, idempotency_key]
              properties:
                agent_id:
                  type: string
                  format: uuid
                amount_minor:
                  type: integer
                  format: int64
                second_authoriser_id:
                  type: string
                  format: uuid
                  description: Must differ from caller (four-eyes)
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Settlement posted
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      tags: [Branch]
      summary: List settlements for the caller's region
      operationId: listSettlements
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, CONFIRMED]
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryOffset"
      responses:
        "200":
          description: Settlement list
  /branch/disputes:
    get:
      tags: [Branch]
      summary: List open disputes in the caller's region
      operationId: listDisputes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryOffset"
      responses:
        "200":
          description: Dispute list
  /branch/disputes/{id}/resolve:
    post:
      tags: [Branch]
      summary: Resolve a transaction dispute (four-eyes)
      operationId: resolveDispute
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution, second_authoriser_id]
              properties:
                resolution:
                  type: string
                  enum: [ACCEPT, REJECT]
                reason:
                  type: string
                  maxLength: 500
                second_authoriser_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Dispute resolved
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT access token. Obtain via /identity/login or /identity/token/refresh.
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
    QueryFrom:
      in: query
      name: from
      schema:
        type: string
        format: date
      description: Filter from date (inclusive)
    QueryTo:
      in: query
      name: to
      schema:
        type: string
        format: date
      description: Filter to date (inclusive)
    QueryLimit:
      in: query
      name: limit
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    QueryOffset:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "INSUFFICIENT_BALANCE"
            message:
              type: string
              description: Human-readable message
              example: "Your balance is too low for this transaction"
        meta:
          $ref: "#/components/schemas/MetaBlock"
      required: [error]
    MetaBlock:
      type: object
      properties:
        request_id:
          type: string
          example: "req_01J3ABCDEF"
        trace_id:
          type: string
          example: "trace_01J3ABCDEF"
...
  responses:
...
    Forbidden:
      description: Insufficient role/permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "FORBIDDEN"
              message: "This action requires BRANCH_MANAGER role"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
...